# runtime/openapi.json
{
  "info": {
    "description": "Runtime HTTP API (served by orbit-tauri-tools runtime).",
    "title": "Orbit Runtime API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/v1/dicts": {
      "get": {
        "description": "Use ?type=xxx to filter; or call /v1/dicts/{type} and /v1/dicts/{type}/{label}.",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "items": [
                    {
                      "id": 1,
                      "label": "RSS",
                      "type": "plugin_type",
                      "value": "rss"
                    }
                  ]
                },
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List dict items",
        "tags": [
          "Dicts"
        ]
      }
    },
...
    "/v1/feed": {
      "get": {
        "description": "Does not fetch remote feeds. Background scheduler refreshes stale plugins per refreshInterval.",
        "parameters": [
          {
            "in": "query",
            "name": "plugin_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "count": 0,
                  "items": [],
                  "ok": true
                },
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get feed items from SQLite cache (read-only)",
        "tags": [
          "Feed"
        ]
      }
    },
    "/v1/feed/item": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plugin_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "channel_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "item": {
                    "id": "1",
                    "title": "Example"
                  },
                  "ok": true
                },
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get a single feed item",
        "tags": [
          "Feed"
        ]
      }
    },
    "/v1/feed/read": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "channelId": "main",
                "id": "native:1",
                "pluginId": "demo"
              },
              "schema": {
                "properties": {
                  "channelId": {
                    "type": "string"
                  },
                  "id": {
                    "type": "string"
                  },
                  "pluginId": {
                    "type": "string"
                  }
                },
                "required": [
                  "id"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "id": "native:1",
                  "ok": true
                },
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Mark feed item as read",
        "tags": [
          "Feed"
        ]
      }
    },
    "/v1/feed/refresh": {
      "post": {
        "parameters": [
          {
            "in": "query",
            "name": "plugin_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Clear cached feed items for the plugin before fetching",
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "count": 0,
                  "ok": true
                },
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Refresh feed for a plugin",
        "tags": [
          "Feed"
        ]
      }
    },
    "/v1/feed/unread": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "plugin_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "channel",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plugin_ids",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "ok": true,
                  "unreadTotal": 0
                },
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Count unread feed items",
        "tags": [
          "Feed"
        ]
      }
    },
    "/v1/images/proxy": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "ok": true
                },
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Proxy image by URL (with SSRF protection)",
        "tags": [
          "Images"
        ]
      }
    },
    "/v1/images/upload": {
      "post": {
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "data": {
                    "image": {
                      "deleteUrl": "https://...",
                      "displayUrl": "https://...",
                      "url": "https://i.imgbb.com/..."
                    }
                  },
                  "ok": true
                },
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Upload image file",
        "tags": [
          "Images"
        ]
      }
    },
    "/v1/images/upload-url": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "url": "https://example.com/image.png"
              },
              "schema": {
                "properties": {
                  "url": {
                    "type": "string"
                  }
                },
                "required": [
                  "url"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "data": {
                    "image": {
                      "deleteUrl": "https://...",
                      "displayUrl": "https://...",
                      "url": "https://i.imgbb.com/..."
                    }
                  },
                  "ok": true
                },
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Upload image by URL",
        "tags": [
          "Images"
        ]
      }
    },
    "/v1/llm/chat/stream": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "messages": [
                  {
                    "content": "Hello!",
                    "role": "user"
                  }
                ],
                "modelId": "gpt-4o-mini",
                "providerId": "openai",
                "stream": true
              },
              "schema": {
                "properties": {
                  "messages": {
                    "items": {
                      "properties": {
                        "content": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "modelId": {
                    "type": "string"
                  },
                  "providerId": {
                    "type": "string"
                  },
                  "stream": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "providerId",
                  "modelId",
                  "messages"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "SSE stream; each `data` payload contains JSON: {delta?:string, done?:bool, error?:string}"
          }
        },
        "summary": "LLM chat (streaming SSE, OpenAI-compatible proxy)",
        "tags": [
          "LLM"
        ]
      }
    },
    "/v1/plugin-groups": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "assignments": {},
                  "collapsed": {},
                  "groups": []
                },
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get plugin groups snapshot",
        "tags": [
          "Plugin Groups"
        ]
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "assignments": {
                  "demo": "default"
                },
                "collapsed": {
                  "default": false
                },
                "groups": [
                  {
                    "id": "default",
                    "label": "默认"
                  }
                ]
              },
              "schema": {
                "properties": {
                  "assignments": {
                    "additionalProperties": {
                      "type": "string"
                    },
                    "type": "object"
                  },
                  "collapsed": {
                    "additionalProperties": {
                      "type": "boolean"
                    },
                    "type": "object"
                  },
                  "groups": {
                    "items": {
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "label": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "type": "array"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "assignments": {},
                  "collapsed": {},
                  "groups": []
                },
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Save plugin groups snapshot",
        "tags": [
          "Plugin Groups"
        ]
      }
    },
    "/v1/plugins": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "plugins": [
                    {
                      "active": true,
                      "icon": "rss",
                      "id": "demo",
                      "name": "Demo"
                    }
                  ]
                },
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List plugins",
        "tags": [
          "Plugins"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "feedUrl": "https://example.com/feed.xml",
                "id": "example",
                "mediaType": "application/rss+xml",
                "name": "Example Feed",
                "refreshInterval": 60,
                "source": "rss"
              },
              "schema": {
                "properties": {
                  "categoryTag": {
                    "type": "string"
                  },
                  "color": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "feedUrl": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  },
                  "id": {
                    "type": "string"
                  },
                  "logoImageUrl": {
                    "type": "string"
                  },
                  "logoText": {
                    "type": "string"
                  },
                  "marketCategory": {
                    "type": "string"
                  },
                  "mediaType": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "refreshInterval": {
                    "type": "integer"
                  },
                  "source": {
                    "type": "string"
                  },
                  "userAgent": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "plugin": {
                    "id": "example"
                  }
                },
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Install plugin (RSS only)",
        "tags": [
          "Plugins"
        ]
      }
    },
...
    "/v1/plugins/order": {
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "orderedIds": [
                  "plugin-a",
                  "plugin-b"
                ]
              },
              "schema": {
                "properties": {
                  "orderedIds": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  }
                },
                "required": [
                  "orderedIds"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "ok": true
                },
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Reorder plugins",
        "tags": [
          "Plugins"
        ]
      }
    },
...
  },
...
}